+++ atlasdb-client/src/main/java/com/palantir/atlasdb/transaction/impl/ForwardingTransaction.java
@@ -96,9 +96,9 @@ public abstract class ForwardingTransaction extends ForwardingObject implements 
     }
 
     @Override
-    public Stream<BatchingVisitable<RowResult<byte[]>>> getUnfetchedRanges(
+    public Stream<BatchingVisitable<RowResult<byte[]>>> getRangesLazy(
             final TableReference tableRef, Iterable<RangeRequest> rangeRequests) {
-        return delegate().getUnfetchedRanges(tableRef, rangeRequests);
+        return delegate().getRangesLazy(tableRef, rangeRequests);
     }
 
     @Override