+++ atlasdb-client/src/test/java/com/palantir/atlasdb/transaction/impl/ReadTransactionShould.java
@@ -58,7 +58,7 @@ public class ReadTransactionShould {
             .put("getRows", new Object[] {DUMMY_THOROUGH_TABLE, ImmutableList.of(EMPTY_BYTES), ColumnSelection.all()})
             .put("getRange", new Object[] {DUMMY_THOROUGH_TABLE, RangeRequest.all()})
             .put("getRanges", new Object[] {DUMMY_THOROUGH_TABLE, ImmutableList.of(RangeRequest.all())})
-            .put("getUnfetchedRanges", new Object[] {DUMMY_THOROUGH_TABLE, ImmutableList.of(RangeRequest.all())})
+            .put("getRangesLazy", new Object[] {DUMMY_THOROUGH_TABLE, ImmutableList.of(RangeRequest.all())})
             .build();
 
     private ReadTransaction readTransaction;
@@ -106,6 +106,7 @@ public class ReadTransactionShould {
         Method[] declaredMethods = ReadTransaction.class.getDeclaredMethods();
 
         for (Method m : declaredMethods) {
+            // Ignore methods that are either not simple gets or are overloaded
             if (simpleGets.containsKey(m.getName()) && simpleGets.get(m.getName()).length == m.getParameterCount()) {
                 checkThrowsAndNoInteraction(() -> {
                             try {