+++ docs/source/release_notes/release-notes.rst
@@ -57,8 +57,11 @@ develop
            Please consult the documentation for :ref:`AtlasDB Configuration <atlas-config>` for details on how to set this up.
            (`Pull Request <https://github.com/palantir/atlasdb/pull/2263>`__)
 
-    *    - |improved|
-         - Added two new methods to Transaction, getUnfetchedRanges and a concurrent version of getRanges. These allow for better control over concurrency than the old getRanges method.
+    *    - |devbreak|
+         - Added two new methods to Transaction, getRangesLazy and a concurrent version of getRanges. The old getRanges method would eagerly load the first page of all ranges, potentially
+           concurrently. This often caused more data to be fetched than necessary or higher concurrency than expected. If you expect to only use a small amount of the rows in the provided
+           ranges, it is often advisable to use the getRangesLazy method and serially iterate over the results. Otherwise, you should use the new version of getRanges that allows explicitly
+           operating on the resulting visitables in parallel.
            (`Pull Request <https://github.com/palantir/atlasdb/pull/2281>`__)
 
     *    - |deprecated|